# python/openapi/generated/server/models/pet_status.py
# coding: utf-8

"""
    Petstore

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.13
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class PetStatus(str, Enum):
    """
    PetStatus
    """

    """
    allowed enum values
    """
    AVAILABLE = 'available'
    PENDING = 'pending'
    SOLD = 'sold'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of PetStatus from a JSON string"""
        return cls(json.loads(json_str))